+++ lib/coderay/styles/style.rb
@@ -1,11 +1,6 @@
 module CodeRay
-
-  # This module holds the Style class and its subclasses.
-  # 
-  # See Plugin.
+  
   module Styles
-    extend PluginHost
-    plugin_path File.dirname(__FILE__), 'styles'
     
     # Base class for styles.
     #
