+++ lib/coderay/tokens.rb
@@ -67,7 +67,7 @@ module CodeRay
     def method_missing meth, options = {}
       encode meth, options
     rescue PluginHost::PluginNotFound
-      super
+      raise
     end
     
     # Split the tokens into parts of the given +sizes+.
